SLING-7811 - NPE when repository is starting up
Make the repository initialisation and startup process more robust
against restarts.

It is possible that while a the repository startup is happening in an
async manner in the "Apache Sling Repository Startup Thread" a stop is
requested. This thread should be notified of any stop requests and
gracefully handle the shutdown.